python/pyproject.toml
[tool.poetry]
name = "karlson_wishlist"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
bs4 = "^0.0.1"
requests = "^2.25.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
